fix: move products views to product directory
