+++ src/views/products/ProductDetail.vue
